requirements.txt
# ============================================
# SYMYAH - Requisitos Otimizados para Produção
# ============================================

# --- Web Framework & API ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# --- Segurança & Autenticação ---
python-jose[cryptography]==3.3.4
passlib[argon2]==1.7.4
PyJWT==2.8.0

# --- Banco de Dados ---
psycopg2-binary==2.9.9
pgvector==0.2.4

# --- Variáveis de Ambiente ---
python-dotenv==1.0.0

# --- IA - OpenAI ---
openai==1.12.0

# --- IA - Google Gemini ---
google-genai==0.2.1

# --- Processamento de Texto ---
nltk==3.8.1

# --- Utilitários ---
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
PyMuPDF==1.23.8

# --- Métricas & Utilitários ---
numpy==1.26.3
scipy==1.12.0
